Notation, spacing, and layout
Up to about line 5930 but then it gets really messy. Lots of standardising of equations and layout. Lots of \left( and \right), etc. to make sure spacing works out for brackets. Been changing arrays to align* environments and removing arbitrary spacings like "\," and "\quad" where unnecessary.

Equations that should be sorted are indented in the .tex file. Same with align*'s.

Might make all of the commutative diagrams the same format. Otherwise the spacing and arrowheads are different all over the place.

Lots of typos, notation, and wording fixed along the way. Lots to do in the later sections to make the tone similar to the rest.